requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.24.0
requests
numpy
pandas
matplotlib
seaborn
pillow
scikit-learn
tqdm
jupyter
kaggle
huggingface_hub
rawpy
pillow_heif
wikipedia